auto add product data by selenium into database and product img into wwwroot success

+++ Models/DiscountRepository.cs
@@ -28,10 +28,11 @@ namespace Shop.Models{
 
         public int Delete(int id){
             Discount dis = GetDiscountById(id);
-            using(dbContext){
+            if(dis != null){
                 dbContext.Discount.Remove((Discount)dis);
                 return dbContext.SaveChanges();
             }
+            return 0;
         }
 
         public int Edit(Discount obj){
